backend/requirements.txt
django
djangorestframework
django-cors-headers
pyjwt
psycopg2-binary
